refactor(facility): 편의시설 조회 관련 위젯 추출
