+++ lib/screens/nearby_facility_screen.dart
@@ -1,12 +1,7 @@
-import 'package:apple_maps_flutter/apple_maps_flutter.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
-import 'package:geolocator/geolocator.dart';
-import 'package:intl/intl.dart';
-import 'package:wooyoungsoo/models/facility_model.dart';
-import 'package:wooyoungsoo/services/facility_service/facility_service.dart';
 import 'package:wooyoungsoo/utils/constants.dart';
 import 'package:wooyoungsoo/widgets/common/navigation_bar_widget.dart';
+import 'package:wooyoungsoo/widgets/nearby_facility_screen/facility_map_widget.dart';
 
 /// 주변 편의시설 화면
 class NearbyFacilityScreen extends StatefulWidget {
@@ -47,370 +42,6 @@ class _NearbyFacilityScreenState extends State<NearbyFacilityScreen> {
       bottomNavigationBar:
           const PetdoriNavigationBar(currentIndex: currentIndex),
       body: FacilityMap(screenWidth: screenWidth, screenHeight: screenHeight),
-    );
-  }
-}
-
-class FacilityMap extends StatefulWidget {
-  const FacilityMap({
-    Key? key,
-    required this.screenWidth,
-    required this.screenHeight,
-  }) : super(key: key);
-  final double screenWidth;
-  final double screenHeight;
-
-  @override
-  State<FacilityMap> createState() => _FacilityMapState();
-}
-
-class _FacilityMapState extends State<FacilityMap> {
-  final FacilityService facilityService = FacilityService();
-  List<FacilityModel> nearbyFacilities = [];
-  late AppleMapController mapController;
-  late Position currentPosition;
-  BitmapDescriptor? hospitalIcon, beautyIcon, hotelIcon;
-  bool isLoading = true;
-
-  void _onMapCreated(AppleMapController controller) {
-    mapController = controller;
-  }
-
-  void loadIcons() async {
-    hospitalIcon = await BitmapDescriptor.fromAssetImage(
-      const ImageConfiguration(),
-      "assets/images/hospital_icon.png",
-    );
-    beautyIcon = await BitmapDescriptor.fromAssetImage(
-      const ImageConfiguration(),
-      "assets/images/beauty_icon.png",
-    );
-    hotelIcon = await BitmapDescriptor.fromAssetImage(
-      const ImageConfiguration(),
-      "assets/images/hotel_icon.png",
-    );
-    setState(() {
-      isLoading = false;
-    });
-  }
-
-  void getCurrentPosition() async {
-    currentPosition = await Geolocator.getCurrentPosition();
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    getCurrentPosition();
-    loadIcons();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Expanded(
-          child: isLoading
-              ? const SizedBox(
-                  height: 90,
-                  child: SpinKitCircle(
-                    color: mainColor,
-                    size: 50,
-                  ),
-                )
-              : Stack(
-                  children: [
-                    AppleMap(
-                      annotations: Set<Annotation>.of(
-                        nearbyFacilities.map(
-                          (facility) => Annotation(
-                            annotationId: AnnotationId(facility.name),
-                            position:
-                                LatLng(facility.latitude, facility.longitude),
-                            icon: hospitalIcon!,
-                            onTap: () {
-                              showBottomSheet(
-                                context: context,
-                                backgroundColor: transparentColor,
-                                builder: (BuildContext context) {
-                                  return FacilityBottomSheet(
-                                    facility: facility,
-                                    width: widget.screenWidth,
-                                    height: widget.screenHeight * 0.18,
-                                  );
-                                },
-                              );
-                            },
-                          ),
-                        ),
-                      ),
-                      onMapCreated: _onMapCreated,
-                      trackingMode: TrackingMode.follow,
-                      initialCameraPosition: CameraPosition(
-                        target: LatLng(currentPosition.latitude,
-                            currentPosition.longitude),
-                      ),
-                    ),
-                    Positioned(
-                      left: widget.screenWidth * 0.5 - 80,
-                      child: ElevatedButton(
-                        style: ElevatedButton.styleFrom(
-                          fixedSize: const Size(160, 36),
-                          backgroundColor: whiteColor,
-                          foregroundColor: mainColor,
-                          shadowColor: Colors.transparent,
-                          shape: RoundedRectangleBorder(
-                            side: const BorderSide(
-                              color: mediumGreyColor,
-                              width: 0.5,
-                            ),
-                            borderRadius: BorderRadius.circular(100),
-                          ),
-                        ),
-                        onPressed: () async {
-                          await mapController.getZoomLevel();
-                          LatLngBounds res =
-                              await mapController.getVisibleRegion();
-                          LatLng northEast = res.northeast;
-                          LatLng southWest = res.southwest;
-                          double centerLat =
-                              (northEast.latitude + southWest.latitude) / 2;
-                          double centerLng =
-                              (northEast.longitude + southWest.longitude) / 2;
-
-                          int distance = Geolocator.distanceBetween(centerLat,
-                                  centerLng, centerLat, southWest.longitude)
-                              .toInt();
-                          print(
-                              "${northEast.latitude}, ${northEast.longitude}");
-                          print("centerLat: $centerLat, centerLng: $centerLng");
-                          print("distance: $distance");
-
-                          nearbyFacilities =
-                              await facilityService.getNearbyFacilities(
-                            latitude: centerLat,
-                            longitude: centerLng,
-                            radius: distance,
-                          );
-
-                          setState(() {});
-                        },
-                        child: const Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            Icon(
-                              Icons.refresh_rounded,
-                              size: 18,
-                              color: mainColor,
-                            ),
-                            SizedBox(
-                              width: 5,
-                            ),
-                            Baseline(
-                              baselineType: TextBaseline.alphabetic,
-                              baseline: 14,
-                              child: Text(
-                                "현 지도에서 검색",
-                                style: TextStyle(
-                                  fontSize: 14,
-                                  fontWeight: fontWeightMedium,
-                                  color: mainColor,
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    Positioned(
-                      right: 0,
-                      top: widget.screenHeight * 0.5,
-                      child: ElevatedButton(
-                        style: ElevatedButton.styleFrom(
-                            fixedSize: const Size(40, 40),
-                            backgroundColor: whiteColor,
-                            foregroundColor: mainColor,
-                            shadowColor: Colors.transparent,
-                            padding: EdgeInsets.zero,
-                            shape: const CircleBorder()),
-                        onPressed: () {
-                          getCurrentPosition();
-                          mapController
-                              .animateCamera(CameraUpdate.newLatLng(LatLng(
-                            currentPosition.latitude,
-                            currentPosition.longitude,
-                          )));
-                        },
-                        child: const Icon(
-                          Icons.gps_fixed_rounded,
-                          size: 20,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-        ),
-      ],
-    );
-  }
-}
-
-class FacilityBottomSheet extends StatelessWidget {
-  FacilityBottomSheet({
-    Key? key,
-    required this.facility,
-    required this.width,
-    required this.height,
-  }) : super(key: key);
-
-  final FacilityModel facility;
-  final double width, height;
-  final Map<String, String> dayOfTheWeek = {
-    "Sun": "일",
-    "Mon": "월",
-    "Tue": "화",
-    "Wed": "수",
-    "Thu": "목",
-    "Fri": "금",
-    "Sat": "토",
-  };
-
-  String getCurrentOperatingInfo(String operatingHourInfo) {
-    List<String> operatingDayOfTheWeeks =
-        operatingHourInfo.split(" ")[0].split(",");
-    String openHour = operatingHourInfo.split(" ")[1];
-    String closeHour = operatingHourInfo.split(" ")[3];
-
-    var now = DateTime.now();
-    String? currentDayOfTheWeek = dayOfTheWeek[DateFormat.E().format(now)];
-
-    if (currentDayOfTheWeek != null &&
-        operatingDayOfTheWeeks.contains(currentDayOfTheWeek)) {
-      DateTime openTime = DateTime(now.year, now.month, now.day,
-          int.parse(openHour.split(":")[0]), int.parse(openHour.split(":")[1]));
-      DateTime closeTime = DateTime(
-          now.year,
-          now.month,
-          now.day,
-          int.parse(closeHour.split(":")[0]),
-          int.parse(closeHour.split(":")[1]));
-      if (now.isAfter(openTime) && now.isBefore(closeTime)) {
-        return "영업 중";
-      }
-    }
-    return "영업 종료";
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: height,
-      width: width,
-      decoration: const BoxDecoration(
-        color: whiteColor,
-        borderRadius: BorderRadius.only(
-          topLeft: Radius.circular(30),
-          topRight: Radius.circular(30),
-        ),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.all(20),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              facility.name,
-              style: const TextStyle(
-                fontSize: 20,
-                fontWeight: fontWeightBold,
-                color: blackColor,
-              ),
-            ),
-            const SizedBox(
-              height: 10,
-            ),
-            facility.operatingHourInfo.isEmpty
-                ? const SizedBox.shrink()
-                : Row(
-                    children: [
-                      SizedBox(
-                        width: 70,
-                        child: Text(
-                          getCurrentOperatingInfo(facility.operatingHourInfo),
-                          style: const TextStyle(
-                            fontSize: 14,
-                            fontWeight: fontWeightBold,
-                            color: redColor,
-                          ),
-                        ),
-                      ),
-                      Text(
-                        facility.operatingHourInfo,
-                        style: const TextStyle(
-                          fontSize: 14,
-                          fontWeight: fontWeightRegular,
-                          color: blackColor,
-                        ),
-                      ),
-                    ],
-                  ),
-            facility.operatingHourInfo.isEmpty
-                ? const SizedBox.shrink()
-                : const SizedBox(
-                    height: 10,
-                  ),
-            Row(
-              children: [
-                SizedBox(
-                  width: 70,
-                  child: Text(
-                    facility.distanceInfo,
-                    style: const TextStyle(
-                      color: blackColor,
-                      fontSize: 14,
-                      fontWeight: fontWeightBold,
-                    ),
-                  ),
-                ),
-                Expanded(
-                  child: Text(
-                    facility.address,
-                    style: const TextStyle(
-                      color: darkGreyColor,
-                      fontSize: 14,
-                      fontWeight: fontWeightRegular,
-                    ),
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ),
-                const SizedBox(width: 10),
-                TextButton.icon(
-                  style: TextButton.styleFrom(
-                    minimumSize: Size.zero,
-                    padding: EdgeInsets.zero,
-                    tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                    shadowColor: Colors.transparent,
-                  ),
-                  onPressed: () {},
-                  icon: const Icon(
-                    Icons.copy_rounded,
-                    color: mainColor,
-                    size: 18,
-                  ),
-                  label: const Text(
-                    "복사",
-                    style: TextStyle(
-                      color: mainColor,
-                      fontSize: 14,
-                      fontWeight: fontWeightMedium,
-                    ),
-                  ),
-                )
-              ],
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
